Move the versioning dataclass and typedict to a dedicated file
